Add files via upload
This keylogger is for education/research purposes only. The author takes NO responsibility and/or liability for how you choose to use any of the source code/any files provided. The author and anyone affiliated with will not be liable for any losses and/or damages in connection with use of ANY files provided with this program. By using this program or any files included, you understand that you are AGREEING TO USE AT YOUR OWN RISK. Once again this file and ALL files included are for EDUCATION and/or RESEARCH purposes ONLY. This file is ONLY intended to be used on your own pentesting labs, or with explicit consent from the owner of the property being tested